% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{build_grid}
\alias{build_grid}
\title{Build a grid for grid search CV.}
\usage{
build_grid(
  times = 30,
  h = 30,
  n_closest = 20,
  offset = rlang::zap(),
  window_args = list(),
  sampler_args = list(),
  train_test_split_type = rlang::zap()
)
}
\arguments{
\item{times}{The number of simulated sample paths to produce per hot deck forecast.}

\item{h}{How many days to forecast.}

\item{n_closest}{vector of scalars or list of not-all-scalars.
The number of closest observations to pick from
per hot deck random sampling. Either scalar (length == 1) or vector
of length == h.}

\item{offset}{integer. Offset (in +- days) from the most recent row of .data
to use as the starting point.}

\item{window_args}{list of \code{build_window_args} outputs.}

\item{sampler_args}{list of \code{build_sampler_args} outputs.}

\item{train_test_split_type}{See \code{cv_hot_deck_forecast} details for details.}
}
\value{
tibble. Grid of parameters for use in grid search.
}
\description{
The output tibble is for use in \code{grid_search_hot_deck_cv} as the \code{grid}
argument.
}
\details{
If you want to pass a non-scalar to the grid search, you need to wrap
the argument in \code{list()}. e.g. \verb{n_closest = list(30, c(rep(20,10),19:15)}
or \code{window_args = list(build_window_args(3))}. Ergo, you \emph{always} need
to wrap any \code{window_args} or \code{sampler_args} in a list.
}
