% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{build_sampler_args}
\alias{build_sampler_args}
\title{Build grid search sampler argument.}
\usage{
build_sampler_args(sa_name, sampler, appender, cov_fc_getter = NULL)
}
\arguments{
\item{sa_name}{The name of your sampler-appender pair.}

\item{sampler}{The sampler.}

\item{appender}{The appender.}

\item{cov_fc_getter}{The cov_fc_getter.}
}
\value{
list(sa_name = sa_name, sampler = sampler, appender = appender)
}
\description{
Use this function to build the \code{sampler_args} arg of \code{build_grid}.
}
\details{
The \code{sa_name} argument is for downstream usage in performance
assessments as a... name.
}
