% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{calc_cv_crps}
\alias{calc_cv_crps}
\title{Calculate the CRPS from hot deck CV.}
\usage{
calc_cv_crps(cv_out, obs_col_name)
}
\arguments{
\item{cv_out}{Output of \code{cv_hot_deck_forecast}.}

\item{obs_col_name}{string. the observation column name.}
}
\value{
\code{cv_out$test_data_sets}, augmented with the column \code{score},
which is the CRPS of the simulated values against the given observation.
}
\description{
Augments the CV's test data with a "score" column of CRPSes, calculated
by \code{scoringRules::crps_sample}.
}
\details{
The score is listed as NA for missing test data observations.

The CRPS uses the \code{times} (arg passed to \code{hot_deck_forecast}
via \code{cv_hot_deck_forecast}) number of simulated values, e.g. if
\code{times = 30} then the 30 forecasted values, for each row in the test data.
}
