% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{cv_hot_deck_forecast}
\alias{cv_hot_deck_forecast}
\title{Generate a num-seasons-fold(-ish) hot deck forecast.}
\usage{
cv_hot_deck_forecast(
  .data,
  .datetime,
  .observation,
  offset,
  times,
  h,
  window_back,
  window_fwd,
  n_closest
)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{.observation}{The observation column of .data. Passed via pipe.}

\item{offset}{integer. Offset (in +- days) from the most recent row of .data
to use as the starting point.}

\item{times}{The number of simulated sample paths to produce per hot deck forecast.}

\item{h}{How many days to forecast.}

\item{window_back}{How many days back to include in the window for
a given season.}

\item{window_fwd}{How many days forward to include in the window for
a given season.}

\item{n_closest}{The number of closest observations to pick from
per hot deck random sampling.}
}
\value{
tibble of hot deck forecasts:
\itemize{
\item nrow = (h * times) \* k (k is # of viable seasons),
\item columns:
\itemize{
\item datetime: the date for the forecast
\item forecast: the forecasted value
\item simulation_num: the simulated sample path number
\item k: the CV number, i.e. the hot deck forecast number
}
}
}
\description{
Produces a set of hot deck forecasts, one per season if possible.
}
\details{
If the \code{offset = 0}, then the CV will use the latest datetime as a
starting point, and go back one year at a time through the data.
If the observation is missing, then that year's hot deck forecast is
skipped. The latest datetime is also skipped, because there's no test
data to use for its forecasts.

If \code{offset} is non-zero, then the starting point will be the latest
datetime + the offset.

The parameters after \code{offset}, i.e. \code{times} through \code{n_closest},
are passed to \code{hot_deck_forecast}.

The current implementation splits the data into training and testing sets
as follows: The testing set is any data in the forecast horizon (\code{h}) for
the current season in question. The training data is any data before
that or after the \code{window_fwd} from the largest \code{h}. This is
a conservative approach, aimed at eliminating data leakage. While the data
after the largest \code{h} but within the \code{window_fwd} should be minimally leaky,
it seemed off for larger \code{h}'s to have increasing pools of data to draw from.

I think that it might be worthwhile to offer an alternative splitting method,
in which all of the data is in the training set. While this would leak data,
it might be a better option for cases of only a few seasons worth of data.
In that case, the current, conservative implementation would have markedly
reduced data to draw from, which seems like it could significantly effect
the conclusion on which hyperparameter values to select.
}
