% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ensure_vector}
\alias{ensure_vector}
\title{Converts scalar to vector of length h, and raises an error if
vector is not of length h.}
\usage{
ensure_vector(hotdeck_arg, h)
}
\arguments{
\item{hotdeck_arg}{Argument passed to \code{hot_deck_forecast} required to end as
a vector of length h.}

\item{h}{Forecast horizon, per \code{hot_deck_forecast}.}
}
\value{
hotdeck_arg vector of length h.
}
\description{
Converts scalar to vector of length h, and raises an error if
vector is not of length h.
}
