% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckfc.R
\name{get_local_rows}
\alias{get_local_rows}
\title{Get rows near (in a seasonal sense) a reference date.}
\usage{
get_local_rows(.data, .datetime, h_curr, window_back, window_fwd)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{h_curr}{Current forecast horizon position.}

\item{window_back}{How many days back to include in the window for
a given season.}

\item{window_fwd}{How many days forward to include in the window for
a given season.}
}
\description{
The reference date is the max date of the data, offset by \code{h_curr}.
So, if your latest data is for 2023-03-03, and ignoring the offset for
simplicity, it collects the rows within a window of March 3 across
all years of your data.
}
