% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{grid_search_build_sampler_args}
\alias{grid_search_build_sampler_args}
\title{Build grid search sampler argument.}
\usage{
grid_search_build_sampler_args(sm_name, sampler, mutator)
}
\arguments{
\item{sm_name}{The name of your sampler-mutator pair.}

\item{sampler}{The sampler.}

\item{mutator}{The mutator.}
}
\value{
list(list(sm_name = sm_name, sampler = sampler, mutator = mutator))
}
\description{
This is a convenience function for if you want to pass
sampler-mutator combinations to \code{grid_search_hot_deck_cv}.
}
\details{
The \code{sm_name} argument is for downstream usage in performance
assessments as a... name.
}
