% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{hot_deck_impute}
\alias{hot_deck_impute}
\title{Hot-deck Impute}
\usage{
hot_deck_impute(
  .data,
  .observation,
  max_gap = 1000,
  n_imputations = 5,
  window_back = 20,
  window_fwd = 20,
  n_closest = 5,
  sampler = c("lead", "diff")
)
}
\arguments{
\item{.data}{tsibble. The data}

\item{.observation}{symbol. The observations column.}

\item{max_gap}{integer. Largest gap size to impute. Larger gaps return NAs.}

\item{n_imputations}{integer. Number of imputations to perform.}

\item{window_back}{integer.
How many days back to include in the window for a given season.}

\item{window_fwd}{integer.
How many days forward to include in the window for a given season.}

\item{n_closest}{integer.
The number of closest observations to pick from per hot deck random sampling.}

\item{sampler}{Sampler function to generate forecasted values.}
}
\value{
.data, augmented with \code{n_imputations} new columns of imputed
observations, named "imputation_{n}".
}
\description{
Impute using hot deck forecasting.
}
\details{
Restrictions on the input data:
\enumerate{
\item The data is a \code{tsibble}.
\item The tsibble is not multi-keyed (i.e. multiple time serieses).
\item The data has no gaps (i.e. missing rows; NA observations okay).
\item The index is a Date.
}

\code{max_gap} and your windows are related. If your window contains no
observations, then the imputation will error. So, you need to make sure
that the \code{max_gap} is small enough -- or that your window is wide enough --
so that there are always observations to sample from.

\code{window_back}, \code{window_fwd}, and \code{n_closest} are restricted to being
scalars for imputation (they can be vectors when calling
\code{\link[=hot_deck_forecast]{hot_deck_forecast()}}).

For \code{sampler}, \code{"diff"} can be tried if \code{"lead"} isn't producing
sensible forecasts. It uses differences between observations -- rather
than taking those adjacent observations -- to generate the imputed
values. As such, it can inject some diversity into sparse regions.
}
