% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Hot-deck Impute}
\usage{
impute(
  .data,
  .observation,
  n_imputations = 5,
  window_back = 20,
  window_fwd = 20,
  n_closest = 5,
  sampler = sample_lead("next_obs")
)
}
\arguments{
\item{.data}{tsibble. The data}

\item{.observation}{symbol. The observations column.}

\item{n_imputations}{integer. Number of imputations to perform.}

\item{window_back}{integer (scalar OR vector).
How many days back to include in the window for a given season.
Either scalar (length == 1) or vector of length == h.}

\item{window_fwd}{integer (scalar OR vector).
How many days forward to include in the window for a given season.
Either scalar (length == 1) or vector of length == h.}

\item{n_closest}{integer (scalar OR vector).
The number of closest observations to pick from per hot deck random sampling.
Either scalar (length == 1) or vector of length == h.}

\item{sampler}{Sampler function to generate forecasted values.}
}
\value{
.data, augmented with \code{n_imputations} new columns of imputed
observations, named "imputation_{n}".
}
\description{
Impute using hot deck forecasting.
}
