% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{internal_basic_hot_deck_sampler}
\alias{internal_basic_hot_deck_sampler}
\title{Wrapped method for \code{basic_hot_deck_sampler}.}
\usage{
internal_basic_hot_deck_sampler(
  local_rows,
  .observation,
  current_obs,
  n_closest,
  next_obs_col_name,
  n_bins
)
}
\arguments{
\item{local_rows}{tibble (NOT tsibble) The \code{local_rows} in \code{simulate_sample_path}.}

\item{.observation}{The observation column.}

\item{current_obs}{The \code{current_obs} in \code{simulate_sample_path}.}

\item{n_closest}{Scalar.}

\item{next_obs_col_name}{Name of the column in your data containing the next
observation (from, e.g. \code{dplyr::lead}).}

\item{n_bins}{Number of bins to use. See details for... details.}
}
\value{
list(new_current_obs, forecast), where
new_current_obs = The value to use for \code{current_obs} in the next iteration,
i.e. the next forecast.
forecast = The forecasted value.
}
\description{
Wrapped method for \code{basic_hot_deck_sampler}.
}
