% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{internal_hot_deck_fc_cov_sam_appl_covs}
\alias{internal_hot_deck_fc_cov_sam_appl_covs}
\title{Covariate sampling intermediary to take in covariate forecasts.}
\usage{
internal_hot_deck_fc_cov_sam_appl_covs(
  covariate_forecast,
  next_target_obs_col_name,
  filter_na_col_names
)
}
\arguments{
\item{covariate_forecast}{tsibble. A simulated sample path forecast
for a covariate.}

\item{next_target_obs_col_name}{The column name of the lead of the target
observation.}

\item{filter_na_col_names}{The column names for which rows with NAs should
be removed before random sampling.}
}
\value{
Partially applied function of the sampling method now with its
covariates.
}
\description{
This is partial application 2 of 2. It is designed to be internal to
\code{hot_deck_forecast}. It takes and processes the covariate forecasts and
returns the final sampling method.
}
