% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckfc.R
\name{internal_hot_deck_forecast}
\alias{internal_hot_deck_forecast}
\title{Internal method for hot deck forecasting.}
\usage{
internal_hot_deck_forecast(
  .data,
  .datetime,
  .observation,
  times,
  h,
  window_back,
  window_fwd,
  n_closest,
  sampler
)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data.}

\item{.observation}{The observation column of .data.}

\item{times}{The number of simulated sample paths to produce.}

\item{h}{How many days to forecast.}

\item{window_back}{How many days back to include in the window for
a given season. Either scalar (length == 1) or vector
of length == h.}

\item{window_fwd}{How many days forward to include in the window for
a given season. Either scalar (length == 1) or vector
of length == h.}

\item{n_closest}{The number of closest observations to pick from
per hot deck random sampling. Either scalar (length == 1) or vector
of length == h.}

\item{sampler}{Sampler function to generate forecasted values.}
}
\description{
Internal method for hot deck forecasting.
}
