% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{internal_hot_deck_n_leads_sampler}
\alias{internal_hot_deck_n_leads_sampler}
\title{Wrapped method for \code{hot_deck_n_leads_sampler}.}
\usage{
internal_hot_deck_n_leads_sampler(
  local_rows,
  .observation,
  current_obs,
  n_closest,
  filter_na_col_names
)
}
\arguments{
\item{local_rows}{tibble (NOT tsibble) The \code{local_rows} in \code{simulate_sample_path}.}

\item{.observation}{The observation column.}

\item{current_obs}{The \code{current_obs} in \code{simulate_sample_path}.}

\item{n_closest}{Scalar.}

\item{filter_na_col_names}{char vec. The names of any columns that you want to
filter any NAs from.}
}
\value{
list(new_current_obs, forecast), where
new_current_obs = list(leads, last_non_na) What to use for \code{current_obs}
in the next iteration, i.e. the next forecast, where
leads       = any leads for the current sample that have yet
to be returned as a forecast
last_non_na = the most recent non-na lead from the current sample
forecast = The forecasted value.
}
\description{
The \code{current_obs} is a scalar for the first forecast, where it's just the
most recent observation. Subsequently, it is a list containing the leads
of the current sample yet to be returned as a forecast, as well as the
most recent non-na forecast for the current sample.
}
