% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{internal_sample_covariate_lead}
\alias{internal_sample_covariate_lead}
\title{Wrapped method for \code{sample_covariate_lead}.}
\usage{
internal_sample_covariate_lead(
  local_rows,
  .observation,
  current_obs,
  n_closest,
  next_cov_obs_col_name,
  next_target_obs_col_name,
  filter_na_col_names
)
}
\arguments{
\item{local_rows}{tibble (NOT tsibble) The \code{local_rows} in \code{simulate_sample_path}.}

\item{.observation}{The covariate's observation column.}

\item{current_obs}{The \code{current_obs} in \code{simulate_sample_path}.}

\item{n_closest}{Scalar.}

\item{next_cov_obs_col_name}{Name of the column in your data
containing the next covariate observation (from, e.g. \code{lead_cov_mutator}).}

\item{next_target_obs_col_name}{The corresponding target observation
(from, e.g. \code{lead_cov_mutator}).}

\item{filter_na_col_names}{char vec. The names of any columns that you want to
filter any NAs from.}
}
\value{
list(new_current_obs, forecast), where
new_current_obs = The value to use for \code{current_obs} in the next iteration,
i.e. the next forecast.
forecast = The forecasted value.
}
\description{
Wrapped method for \code{sample_covariate_lead}.
}
