% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{internal_sample_forecasted_covariate}
\alias{internal_sample_forecasted_covariate}
\title{Actual sampling method for \code{sample_forecasted_covariate}.}
\usage{
internal_sample_forecasted_covariate(
  local_rows,
  .cov_observation,
  current_obs,
  n_closest,
  covariate_forecast,
  next_target_obs_col_name,
  filter_na_col_names
)
}
\arguments{
\item{local_rows}{tibble (NOT tsibble) The \code{local_rows} in \code{simulate_sample_path}.}

\item{.cov_observation}{The covariate's observation column.}

\item{current_obs}{The \code{current_obs} in \code{simulate_sample_path}.}

\item{n_closest}{Scalar.}

\item{covariate_forecast}{list. A simulated sample path forecast for a covariate.}

\item{next_target_obs_col_name}{The lead of the target observation.}

\item{filter_na_col_names}{The column names for which rows with NAs should
be removed before random sampling.}
}
\value{
list(new_current_obs, forecast), where
new_current_obs = list. The covariate's forecasts.
Contains the value to use for \code{current_obs} in the next iteration at index 1.
forecast = The forecasted value.
}
\description{
Actual sampling method for \code{sample_forecasted_covariate}.
}
