% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{lead_cov_mutator}
\alias{lead_cov_mutator}
\title{Add a column of covariate and target leads to data.}
\usage{
lead_cov_mutator(.data, .cov_observation)
}
\arguments{
\item{.data}{The data. A tsibble.}

\item{.cov_observation}{The covariate observation column. Passed via pipe.}
}
\value{
.data, augmented with a column, named \code{next_obs}, of leads.
}
\description{
Adds two new columns, named "next_cov_obs" and "next_target_obs",
of the next (i.e. lead) covariate and target observations.
}
\details{
The corresponding selector, \code{hot_deck_covariate_lead_sampler}, defaults to
these names for the lead columns, so it makes things slightly easier in that regard.

This mutator should be paired with \code{hot_deck_covariate_lead_sampler} for use in CV.
It is applied to the training data after the train-test split,
to avoid data leakage.
}
