% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{n_leads_mutator}
\alias{n_leads_mutator}
\title{Add \code{n} columns of leads to data.}
\usage{
n_leads_mutator(.data_ts, .observation, n)
}
\arguments{
\item{.data_ts}{The data. A tsibble.}

\item{.observation}{The observation column. Passed via pipe.}

\item{n}{Number of columns of leads to add.}
}
\value{
.data_ts, augmented with \code{n} new columns, named \code{lead_[1:n]},
of leads.
}
\description{
Adds \code{n} new columns, named \code{lead_[1:n]}, of leads of the observation column.
}
\details{
The corresponding selector is \code{hot_deck_n_leads_sampler}.

This mutator should be paired with \code{hot_deck_n_leads_sampler} for use in CV.
It is applied to the training data after the train-test split,
to avoid data leakage.
}
