% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckfc.R
\name{partially_validate_hotdeckfc_input}
\alias{partially_validate_hotdeckfc_input}
\title{Some validators for hot deck forecast input.}
\usage{
partially_validate_hotdeckfc_input(data)
}
\arguments{
\item{data}{The input \code{.data} to \code{hot_deck_forecast}.}
}
\description{
Does not check if final obs is NA; put that in the code itself
b/c it logically goes w/ \code{current_obs}. Subject to change.
}
