% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cv_forecasts}
\alias{plot_cv_forecasts}
\title{Plot CV forecasts}
\usage{
plot_cv_forecasts(
  data,
  arg_list,
  cv_out,
  observation_col_name,
  window_back,
  window_fwd,
  forecasts_geom
)
}
\arguments{
\item{data}{tsibble. The data.}

\item{arg_list}{list. The \verb{$arg_list} element in grid search output.}

\item{cv_out}{list. The \verb{$cv_out} element in grid search output.}

\item{observation_col_name}{string. The observation column name.}

\item{window_back}{positive integer. How many days back to include
of the .data.}

\item{window_fwd}{positive integer. How many days forward to include
of the .data.}

\item{forecasts_geom}{string. The geom to use for the forecasts on the ggplot}
}
\description{
Plot the k forecasts for a CV.
}
