% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecast over seasonal history}
\usage{
plot_forecast(
  forecast,
  historical_data,
  observation_col_name = "observation",
  filter_past_forecasts = FALSE,
  forecasts_geom = c("line", "point"),
  title = "Forecast",
  x_lab = "Day of Year",
  y_lab = "Observation",
  caption = ggplot2::waiver()
)
}
\arguments{
\item{forecast}{tibble. Output of \code{\link[=hot_deck_forecast]{hot_deck_forecast()}}.}

\item{historical_data}{tsibble. The data.}

\item{observation_col_name}{string. The observation column name.}

\item{filter_past_forecasts}{bool. Filter forecasts before \code{lubdridate::today()}?}

\item{forecasts_geom}{string. The geom to use for the forecasts on the ggplot}

\item{title}{string. The title.}

\item{x_lab}{string. The x-axis label.}

\item{y_lab}{string. The y-axis label.}

\item{caption}{string. The figure caption.}
}
\description{
Plot forecast over the historical seasonal data.
}
\details{
Using a \code{forecasts_geom} of "line" is more useful for less zig-zaggy
forecasts. This generally means small values of \code{n_closest}. Using "point"
is good for zig-zaggy forecasts, because in these cases, "line" yeilds a
spiky blob.
}
\examples{
data = append_lead(hotdeckfc::SUGG_temp, observation)
fc = hot_deck_forecast(data,
                       date,
                       observation,
                       times = 5,
                       h = 20,
                       window_back = 20,
                       window_fwd = 20,
                       n_closest = 5)
plot_forecast(fc, data)

}
