% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_grid_search_crps}
\alias{plot_grid_search_crps}
\title{Plot CRPSes of grid search}
\usage{
plot_grid_search_crps(
  grid_out,
  observation_col_name = "observation",
  ymax = NULL
)
}
\arguments{
\item{grid_out}{output from \code{\link[=grid_search_hot_deck_cv]{grid_search_hot_deck_cv()}}.}

\item{observation_col_name}{string. The observation column name from your data.}

\item{ymax}{optional numeric. The plots' y max.}
}
\description{
Plot the CRPSes of grid search output along h.
}
\details{
The title of the plot is the various parameter values, where
\itemize{
\item "wf" = window_forward
\item "wb" = window_back
\item "nc" = n_closest
\item "sa_name" = sa_name
For the first wf, wb, and nc, if you provided a vector, then the
sum of that vector is used as the value.
}
}
\examples{
grid = build_grid(
  h = 2,
  n_closest = c(5, 10),
  # even if you only use a single `build_window_args()`, wrap in `list()`
  window_args = list(build_window_args(20)),
  # also need to wrap in list
  sampler_args = list(
    build_sampler_args(sa_name = "nxt",
                       sampler = sample_lead(),  # remember to call!
                       appender = append_lead)
  )
)
out = grid_search_hot_deck_cv(hotdeckts::SUGG_temp,
                              .observation = observation,
                              grid = grid)
plot_grid_search_crps(out, "observation", 5)

}
