% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{plot_imputation}
\alias{plot_imputation}
\title{Plot imputation}
\usage{
plot_imputation(.imputation, .observation)
}
\arguments{
\item{.imputation}{Output of \code{\link[=impute]{impute()}}.}

\item{.observation}{symbol. The observations column.}
}
\description{
Plot output of \code{\link[=impute]{impute()}}
}
