% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impute.R
\name{plot_imputation_gg_season}
\alias{plot_imputation_gg_season}
\title{Plot imputations with \code{\link[feasts:gg_season]{feasts::gg_season()}}.}
\usage{
plot_imputation_gg_season(.imputation, .observation)
}
\arguments{
\item{.imputation}{Output of \code{\link[=hot_deck_impute]{hot_deck_impute()}}.}

\item{.observation}{symbol. The observations column.}
}
\description{
Plot output of \code{\link[=hot_deck_impute]{hot_deck_impute()}} using \code{\link[feasts:gg_season]{feasts::gg_season()}}.
}
\details{
Each imputation, plus the original observations, are plotted
separately (but in one figure) using \code{\link[feasts:gg_season]{feasts::gg_season()}}.
}
\examples{
3
# imputed = hot_deck_impute(hotdeckts::SUGG_temp \%>\%
#   tsibble::as_tsibble(), observation)
# plot_imputation_gg_season(imputed, observation)

}
