% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impute.R
\name{plot_imputation_seasonal_facet}
\alias{plot_imputation_seasonal_facet}
\title{Plot imputations with seasonal facets}
\usage{
plot_imputation_seasonal_facet(.imputation, .observation)
}
\arguments{
\item{.imputation}{Output of \code{\link[=hot_deck_impute]{hot_deck_impute()}}.}

\item{.observation}{symbol. The observations column.}
}
\description{
Plot output of \code{\link[=hot_deck_impute]{hot_deck_impute()}}, faceted by year.
}
\details{
This plotting function spreads the data out more, compared to
\code{\link[=plot_imputation]{plot_imputation()}}, so it's easier to
get a feel for how the imputations look. The imputed sections
are the divergent colored lines, while non-imputed areas (and also any
areas that happen to have all identical imputations) are purple-ish.
}
\examples{
3
# imputed = hot_deck_impute(hotdeckts::SUGG_temp \%>\%
#   tsibble::as_tsibble(), observation)
# plot_imputation_seasonal_facet(imputed, observation)

}
