% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impute.R
\name{plot_imputation_seasonal_separate}
\alias{plot_imputation_seasonal_separate}
\title{Plot imputations seasonally separately}
\usage{
plot_imputation_seasonal_separate(
  .imputation,
  .observation,
  ymin = 0,
  ymax = NULL
)
}
\arguments{
\item{.imputation}{Output of \code{\link[=hot_deck_impute]{hot_deck_impute()}}.}

\item{.observation}{symbol. The observations column.}

\item{ymin}{optional integer. ymin of the plots.}

\item{ymax}{optional integer. ymax of the plots.}
}
\description{
Plot output of \code{\link[=hot_deck_impute]{hot_deck_impute()}}, one figure per year.
}
\details{
This plotting function spreads the data out even more than
\code{\link[=plot_imputation_seasonal_facet]{plot_imputation_seasonal_facet()}}, compared to \code{\link[=plot_imputation]{plot_imputation()}},
The imputed sections
are the divergent colored lines, while non-imputed areas (and also any
areas that happen to have all identical imputations) are purple-ish.
}
\examples{
3
# imputed = hot_deck_impute(hotdeckts::SUGG_temp \%>\%
#   tsibble::as_tsibble(), observation)
# plot_imputation_seasonal_separate(imputed, observation)

}
