% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_covariate_lead}
\alias{sample_covariate_lead}
\title{Sampler of local covariate's leads}
\usage{
sample_covariate_lead(
  next_cov_obs_col_name = "next_cov_obs",
  next_target_obs_col_name = "next_target_obs",
  filter_na_col_names = next_cov_obs_col_name
)
}
\arguments{
\item{next_cov_obs_col_name}{string. Name of the column in your data
containing the next covariate observation. See \code{\link[=append_lead_cov_lead]{append_lead_cov_lead()}}.}

\item{next_target_obs_col_name}{string. The corresponding target observation
See \code{\link[=append_lead_cov_lead]{append_lead_cov_lead()}}.}

\item{filter_na_col_names}{char vec. The names of any columns that you want to
filter any NAs from.}
}
\value{
Partially applied function to be passed to \code{\link[=hot_deck_forecast]{hot_deck_forecast()}}.
}
\description{
Sampler argument for \code{\link[=hot_deck_forecast]{hot_deck_forecast()}} to generate a forecast.
}
\details{
Given a \code{local_rows} and \code{current_obs}, it:
\itemize{
\item Finds those rows in \code{local_rows} with a covariate's
observations closest to \code{current_obs}.
\item Takes the \code{n_closest} of those.
\item Randomly samples one of those.
\item Returns:
\itemize{
\item That sample's "tomorrow"s covariate's observation as the next \code{current_obs}
\item That sample's tomorrow's target variable's observation and the \code{forecast}.
}
}

For \code{n_closest}, \code{dplyr::slice_min} is used, and tie values are included.

For \code{filter_na_col_names}, you must include the \code{next_cov_obs_col_name},
or the sampling will break.
}
\examples{
data = dplyr::mutate(hotdeckts::SUGG_temp,
                     cov_obs = observation + rnorm(n = 1, sd = 3))  # fake cov data
data = append_lead_cov_lead(data, cov_obs, "observation")
data = trim_leading_nas(data, observation)
out = hot_deck_forecast(data,
                        .datetime = date,
                        .observation = observation,
                        times = 3,
                        h = 5,
                        window_back = 20,
                        window_fwd = 20,
                        n_closest = 5,
                        sampler = sample_covariate_lead())  # don't forget to call!

}
