% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_lead}
\alias{sample_lead}
\title{Sampler of local observations' leads}
\usage{
sample_lead(next_obs_col_name = "next_obs")
}
\arguments{
\item{next_obs_col_name}{string. Name of the column in your data containing
the next observation. See \code{\link[=append_lead]{append_lead()}}.}
}
\value{
Partially applied function to be passed to \code{\link[=hot_deck_forecast]{hot_deck_forecast()}}.
}
\description{
Sampler argument for \code{\link[=hot_deck_forecast]{hot_deck_forecast()}} to generate a forecast.
}
\details{
Pass this called function (i.e. \code{sampler = sample_lead()},
NOT \code{sampler = sample_lead}) to \code{\link[=hot_deck_forecast]{hot_deck_forecast()}}.

Given a \code{local_rows} and \code{current_obs}, it:
\itemize{
\item Finds those rows in \code{local_rows} with observations closest to \code{current_obs}.
\item Takes the \code{n_closest} of those.
\item Randomly samples one of those.
\item Returns that sample's "tomorrow" observation as both
the next \code{current_obs} and the \code{forecast}.
}

For \code{n_closest}, \code{dplyr::slice_min} is used, and tie values are included.
}
\examples{
data = append_lead(hotdeckts::SUGG_temp, observation)
data = trim_leading_nas(data, observation)
out = hot_deck_forecast(data,
                        .datetime = date,
                        .observation = observation,
                        times = 3,
                        h = 20,
                        window_back = 20,
                        window_fwd = 20,
                        n_closest = 5,
                        sampler = sample_lead())  # don't forget to call!

}
