% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_local_rows}
\alias{sample_local_rows}
\title{Sample a random row from local rows.}
\usage{
sample_local_rows(
  local_rows,
  .observation,
  current_obs,
  n_closest,
  filter_na_col_names = vector(mode = "character")
)
}
\arguments{
\item{local_rows}{tibble (NOT tsibble) The \code{local_rows} in \code{simulate_sample_path}.}

\item{.observation}{The observation column.}

\item{current_obs}{The \code{current_obs} in \code{simulate_sample_path}.}

\item{n_closest}{Scalar.}

\item{filter_na_col_names}{char vec. The names of any columns that you want to
filter any NAs from.}
}
\value{
A row from \code{local_rows}, without the \code{offset} column anymore.
}
\description{
Procedure outline:
\itemize{
\item Filter rows with NAs in derived_col
\item Take the \code{n_closest} closest rows
\item Randomly sample one of these.
}
}
\details{
The \code{filter_na_col_names} may be something like the column of leads or
differences, or it could be the observation column itself, or it could be
nothing. In some cases a value is required. For instance, NAs are removed
from the leads column for the \code{sample_lead} because these values
are used for subsequent horizons as the "current value".

If you leave NAs in your observation column, then they will only be
possibly sampled if your \code{n_closest} exceeds the number of non-NA observations.
That is, they are a last resort.

If you leave NAs in the column that yields your forecasts (and doing so is a
permissible option), then you will wind up with NAs in your forecasts, which
may be desirable from the perspective of yielding weaker confidence in those
regions with less data to draw from.

The returned row has its \code{offset} column removed. This is just easier for
this function; it would not be difficult to include, but I don't see a reason
to include it; its purpose should have been fulfilled within this function.
}
