% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{shiny_visualize_forecast}
\alias{shiny_visualize_forecast}
\title{Shiny forecast plotter}
\usage{
shiny_visualize_forecast(
  .data,
  .observation,
  samplers = list(lead = sample_lead(), diff = sample_diff()),
  covariate_forecasts = NULL,
  title = ggplot2::waiver()
)
}
\arguments{
\item{.data}{tsibble. The data.}

\item{.observation}{symbol. The observation column of \code{.data}.}

\item{samplers}{list. Any samplers you want to have to option to use.}

\item{covariate_forecasts}{optional tsibble.
Simulated sample paths of covariates.}

\item{title}{optional string. Plot title.}
}
\description{
Widget to play around with forecasting parameters.
}
\details{
Currently, this widget doesn't allow for vectors to be passed to those
forecasting parameters that can take vectors (\verb{window_*}, \code{n_closest}).
}
\examples{
data = append_lead(hotdeckts::SUGG_temp, observation)
#shiny_visualize_forecast(data, date, observation)


}
