% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckfc.R
\name{simulate_sample_path}
\alias{simulate_sample_path}
\title{Simulate a sample path for a hot deck forecast.}
\usage{
simulate_sample_path(
  .data,
  .datetime,
  .observation,
  h,
  window_back,
  window_fwd,
  n_closest,
  sampler
)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{.observation}{The observation column of .data. Passed via pipe.}

\item{h}{How many days to forecast.}

\item{window_back}{How many days back to include in the window for
a given season. Vector of length == h.}

\item{window_fwd}{How many days forward to include in the window for
a given season. Vector of length == h.}

\item{n_closest}{The number of closest observations to pick from
per hot deck random sampling. Vector of length == h.}

\item{sampler}{Sampler function to generate forecasted values.}
}
\value{
tibble of forecasts:
\itemize{
\item nrow = h,
\item columns:
\itemize{
\item datetime: the date for the forecast
\item h:        the forecast horizon
\item forecast: the forecasted value
}
}
}
\description{
Produces a single possible sample path of an h-step forecast.
}
\details{
This method is iterative. For iteration i, at the most recent observation, it:
\itemize{
\item Takes all historical observations from a window,
[\code{window_back[[i]]}, \code{window_fwd[[i]]}] (inclusive) around
the same portion of the season.
\itemize{
\item e.g., Jun 30 +- 5 days across all years.
}
\item Takes the \code{n_closest[[i]]} closest observations from these.
\item Randomly selects one of these closest observations.
\item Uses the observation of the day after that one as the forecasted value.
\item Repeats, using this new forecasted value and its respective date as
the new "most recent observation", up to h forecasts.
}
}
