% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_hotdeckfc.R
\name{slow_hot_deck_forecast}
\alias{slow_hot_deck_forecast}
\title{Produce a hot deck forecast.}
\usage{
slow_hot_deck_forecast(
  .data,
  .datetime,
  .observation,
  times,
  h,
  window_back,
  window_fwd,
  n_closest
)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{.observation}{The observation column of .data. Passed via pipe.}

\item{times}{The number of simulated sample paths to produce.}

\item{h}{How many days to forecast.}

\item{window_back}{How many days back to include in the window for
a given season.}

\item{window_fwd}{How many days forward to include in the window for
a given season.}

\item{n_closest}{The number of closest observations to pick from
per hot deck random sampling.}
}
\value{
tibble of forecasts:
\itemize{
\item nrow = h * times,
\item columns:
\itemize{
\item datetime: the date for the forecast
\item h: the forecast horizon
\item forecast: the forecasted value
\item simulation_num: the simulated sample path number
}
}
}
\description{
Produces multiple h-step forecast simulated sample paths.
}
\details{
For \code{times} times, it produces a simulated sample paths. For each
simulation, the process proceeds as follows:
This method is iterative. At the most recent observation, it:
\itemize{
\item Takes all historical observations from a window around
the same portion of the season.
\itemize{
\item e.g., Jun 30 +- 5 days across all years.
}
\item Takes the \code{n_closest} closest observations from these.
\item Randomly selects one of these \code{n_closest} observations.
\item Uses the observation of the day after that one as the forecasted value.
\item Repeats, using this new forecasted value and its respective date as
the new "most recent observation", up to h forecasts.
}
}
