% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{subtract_year}
\alias{subtract_year}
\title{Subtract a year(ish) from a date.}
\usage{
subtract_year(date)
}
\arguments{
\item{date}{A datetime.}
}
\description{
If the date is in a leap year, subtract 366 days, else 365 days,
so it might yield a different day number.
}
