% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{train_test_split}
\alias{train_test_split}
\title{Train test split for hot deck CV.}
\usage{
train_test_split(
  .data,
  .datetime,
  ref_date,
  h,
  window_fwd,
  min_date,
  split_type
)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{ref_date}{The current "now" per the CV.}

\item{h}{How many days to forecast.}

\item{window_fwd}{How many days forward to include in the window for
a given season. Scalar or vector (see \code{hot_deck_forecast}).}

\item{min_date}{The earliest date in \code{.data}.}

\item{split_type}{c("conservative", "leaky"). See \code{cv_hot_deck_forecast}
help for details.}
}
\value{
list(
train_data = train_data fraction of .data,
test_data  = test_data fraction of .data, augmented with:
h = forecast horizon
)
}
\description{
re \code{split_type}:
"conservataive" excludes the test data, plus the data in the window_fwd
region beyond the final date to forecast, from the training data. If
\code{window_fwd} is a vector, it uses the max.
}
\details{
"leaky" uses all the data as training data.
}
