% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{train_test_split_conservative}
\alias{train_test_split_conservative}
\title{Conservative train test split for hot deck CV.}
\usage{
train_test_split_conservative(
  .data,
  .datetime,
  ref_date,
  h,
  window_fwd,
  min_date
)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{ref_date}{The current "now" per the CV.}

\item{h}{How many days to forecast.}

\item{window_fwd}{How many days forward to include in the window for
a given season.}

\item{min_date}{The earliest date in \code{.data}.}
}
\value{
list(train_data = train_data, test_data = test_data)
}
\description{
Excludes the test data, plus the data in the window_fwd region
beyond the final date to forecast.
}
