% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotdeckfc.R
\name{trim_leading_nas}
\alias{trim_leading_nas}
\title{Remove leading rows with NA observations.}
\usage{
trim_leading_nas(.data, .observation)
}
\arguments{
\item{.data}{The data. Passed via pipe.}

\item{.observation}{The observation column of .data. Passed via pipe.}
}
\description{
Use this function before passing your data to \code{hot_deck_forecast},
which requires that the most recent observation exists (is not NA).
}
