% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{validate_cov_fcs}
\alias{validate_cov_fcs}
\title{Validators for covariate forecasts.}
\usage{
validate_cov_fcs(covariate_forecasts)
}
\arguments{
\item{covariate_forecasts}{tsibble. Simulated sample paths of covariates.}
}
\description{
Checks that:
\enumerate{
\item The cov_fcs is a tsibble.
\item The "key" is one column.
\item The "key" is from 1:n, where n = # of simulations.
}
}
