% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{validate_grid_arg_len}
\alias{validate_grid_arg_len}
\title{Validate potentially-length-h grid arguments.}
\usage{
validate_grid_arg_len(grid_arg, h, name)
}
\arguments{
\item{grid_arg}{An argument for grid search.}

\item{h}{The \code{h} grid search argument.}

\item{name}{The name of the argument.}
}
\description{
Works just like the similar validator in \code{hot_deck_forecast},
just catches this problem before you're halfway through your grid search.
}
\details{
Currently this is an issue for \code{n_closest} and \verb{window_\{fwd, back\}}.

Raises an error if the argument is neither length-h nor scalar.
}
