% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{append_nothing}
\alias{append_nothing}
\title{Return data, unmodified.}
\usage{
append_nothing(.data, .observation)
}
\arguments{
\item{.data}{The data. A tsibble.}

\item{.observation}{The observation column. Passed via pipe.}
}
\value{
.data
}
\description{
This appender should be paired with any sampler functions that do not have
data leakage concerns (as with those that rely on, say, \code{lead} or \code{diff}),
for use in CV.
It is applied to the training data after the train-test split.
}
