% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{build_sampler_args}
\alias{build_sampler_args}
\title{Build grid search sampler argument}
\usage{
build_sampler_args(sa_name, sampler, appender, cov_fc_getter = NULL)
}
\arguments{
\item{sa_name}{string. The name of your sampler-appender pair.}

\item{sampler}{called function. The sampler.}

\item{appender}{function. The appender.}

\item{cov_fc_getter}{optional function. The cov_fc_getter.}
}
\value{
list(sa_name = sa_name, sampler = sampler,
appender = appender, cov_fc_getter = cov_fc_getter)
}
\description{
Use this function to build the \code{sampler_args} arg of \code{\link[=build_grid]{build_grid()}}.
}
\details{
The \code{sa_name} argument is for downstream usage in performance
assessments as a... name.
}
\examples{
build_sampler_args(sa_name = "nxt",
                   sampler = sample_lead(),  # remember to call!
                   appender = append_lead)

}
