% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{build_sampler_args}
\alias{build_sampler_args}
\title{Build grid search sampler argument.}
\usage{
build_sampler_args(sm_name, sampler, mutator)
}
\arguments{
\item{sm_name}{The name of your sampler-mutator pair.}

\item{sampler}{The sampler.}

\item{mutator}{The mutator.}
}
\value{
list(sm_name = sm_name, sampler = sampler, mutator = mutator)
}
\description{
Use this function to build the \code{sampler_args} arg of \code{build_grid}.
}
\details{
The \code{sm_name} argument is for downstream usage in performance
assessments as a... name.
}
