% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{cv_hot_deck_forecast}
\alias{cv_hot_deck_forecast}
\title{Cross-validate hot deck forecast.}
\usage{
cv_hot_deck_forecast(
  .data,
  .datetime,
  .observation,
  times,
  h,
  window_back,
  window_fwd,
  n_closest,
  offset = 0,
  sampler = sample_lead("next_obs"),
  appender = append_lead,
  cov_fc_getter = NULL,
  train_test_split_type = c("conservative", "leaky")
)
}
\arguments{
\item{.data}{tsibble. The data.}

\item{.datetime}{tmp}

\item{.observation}{symbol. The observation column of \code{.data}.}

\item{times}{integer. The number of simulated sample paths to produce.}

\item{h}{integer. How many days to forecast.}

\item{window_back}{integer (scalar OR vector).
How many days back to include in the window for a given season.
Either scalar (length == 1) or vector of length == h.}

\item{window_fwd}{integer (scalar OR vector).
How many days forward to include in the window for a given season.
Either scalar (length == 1) or vector of length == h.}

\item{n_closest}{integer (scalar OR vector).
The number of closest observations to pick from per hot deck random sampling.
Either scalar (length == 1) or vector of length == h.}

\item{offset}{integer. Offset (in +- days) from the most recent row of .data
to use as the starting point.}

\item{sampler}{Sampler function to generate forecasted values.}

\item{appender}{Appender function to be applied to training data.}

\item{cov_fc_getter}{fn(training_max_date) -> tsibble.
Getter of any covariate data, taking the max date of the current training
data.}

\item{train_test_split_type}{default = "conservative". See help for details.}
}
\value{
list containing:
forecasts: a \code{tibble} of hot deck forecasts:
- nrow = (h * times) \* k (k is # of viable seasons),
- columns:
- datetime: the date for the forecast
- forecast: the forecasted value
- simulation_num: the simulated sample path number
- k: the CV number, i.e. the hot deck forecast number
test_data_sets: a \code{tsibble} of test data sets from your \code{.data}
- new columns:
- h: the horizon number
- k: the CV number, i.e. the hot deck forecast number
}
\description{
Produces a set of hot deck forecasts, one per season if possible.
}
\details{
If the observation is missing, then that year's hot deck forecast is
skipped. The latest datetime is also skipped, because there's no test
data to use for its forecasts.

If the \code{offset = 0}, then the CV will use the latest datetime as a
starting point, and go back one year at a time through the data.
If \code{offset} is non-zero, then the starting point will be the latest
datetime + the offset.

The \code{appender} function is for cases where the \code{sampler} uses subsequent
observations (e.g. a \code{lead}). While you can simply add your (e.g. \code{lead})
column to your data before passing it to \code{hot_deck_forecast}, you need
to pass your desired mutation function as an argument to CV. It will
be applied after the train-test split to avoid data leakage. For instance,
the \code{lead} of the final training observation before your test data should
be NA, not the first test data observation.
If you do not want any mutation, use \code{append_nothing} as the argument.

The \code{cov_fc_getter} takes the max date of the current training data as its
only argument. The returned value will be used as the \code{covariate_forecasts}
argument for \code{hot_deck_forecast}. You should consider memoizing the
function you use here.

The current implementation has two train-test data splitting methods.
The default is "conservative", which goes as follows:
The testing set is any data in the forecast horizon (\code{h}) for
the current season in question. The training data is any data before
that or after the \code{max(window_fwd)} from the largest \code{h}. This is aimed at
eliminating data leakage. While the data after the largest \code{h} but within
the \code{max(window_fwd)} should be minimally leaky, it seemed off for
larger \code{h}'s to have increasing pools of data to draw from.

The alternative splitting method, "leaky", uses all of the data
in the training set. While this leaks data, it might be a better option
(TBD) for cases of only a few seasons worth of data.
In that case, the default conservative implementation would have markedly
reduced data to draw from, which seems like it could significantly effect
the conclusion on which hyperparameter values to select.
}
\examples{
out = cv_hot_deck_forecast(hotdeckts::SUGG_temp,
                           .datetime = date,
                           .observation = observation,
                           times = 30,
                           h = 20,
                           window_back = 20,
                           window_fwd = 20,
                           n_closest = 5)

}
