% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{diff_mutator}
\alias{diff_mutator}
\title{Add a column of differences to the next observation.}
\usage{
diff_mutator(.data, .observation)
}
\arguments{
\item{.data}{The data. A tsibble.}

\item{.observation}{The observation column. Passed via pipe.}
}
\value{
.data, augmented with a column, named \code{diff_to_next_obs},
of leads of differences.
}
\description{
Adds a new column, named "diff_to_next_obs", of the differences to the
next (i.e. lead) observations.
}
\details{
The corresponding selector, \code{hot_deck_diff_sampler}, defaults to this name
for the diff column, so it makes things slightly easier in that regard.

This mutator should be paired with \code{hot_deck_diff_sampler} for use in CV.
It is applied to the training data after the train-test split,
to avoid data leakage.
}
