% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{grid_search_hot_deck_cv}
\alias{grid_search_hot_deck_cv}
\title{Perform a grid search CV on \code{hot_deck_forecast}.}
\usage{
grid_search_hot_deck_cv(.data, .datetime, .observation, grid, echo = TRUE)
}
\arguments{
\item{.data}{tsibble. The data. Passed via pipe.}

\item{.datetime}{The datetime column of .data. Passed via pipe.}

\item{.observation}{The observation column of .data. Passed via pipe.}

\item{grid}{tibble. Created using \code{build_grid}.}

\item{echo}{Print progress?}
}
\value{
list, containing two items per element:
args: The arguments that the particular CV was called with.
cv_out: The output of \code{cv_hot_deck_forecast} for those args.
}
\description{
Use \code{build_grid} to build the tibble to pass to the \code{grid} argument.
}
\details{
\code{echo} prints progress like "Finished CV 3 of 8.".
}
