% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{lead_cov_mutator}
\alias{lead_cov_mutator}
\title{Add a column of covariate and target leads to data.}
\usage{
lead_cov_mutator(
  .data_ts,
  .cov_observation,
  target_obs_col_name = "observation"
)
}
\arguments{
\item{.data_ts}{The data. A tsibble.}

\item{.cov_observation}{The covariate observation column. Passed via pipe.}

\item{target_obs_col_name}{The target observation column name.}
}
\value{
.data_ts, augmented with two columns:
\code{next_cov_obs}, of covariate leads.
\code{next_target_obs}, of target leads.
}
\description{
Adds two new columns, named "next_cov_obs" and "next_target_obs",
of the next (i.e. lead) covariate and target observations.
}
\details{
The corresponding, covariate-using selectors, \code{sample_covariate_lead}
and \code{sample_forecasted_covariate} default to these name(s)
for the lead column(s) (the latter doesn't actually use the covariate's lead),
so it makes things slightly easier in that regard.

This mutator should be paired with samplers that rely on
covariates, for use in CV. If your target observation column is not named
"observation", you can use \code{purrr::partial} to pass the right name
before passing that output to CV.

It is applied to the training data after the train-test split,
to avoid data leakage.
}
