% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{lead_mutator}
\alias{lead_mutator}
\title{Add a column of leads to data.}
\usage{
lead_mutator(.data, .observation)
}
\arguments{
\item{.data}{The data. A tsibble.}

\item{.observation}{The observation column. Passed via pipe.}
}
\value{
.data, augmented with a column, named \code{next_obs}, of leads.
}
\description{
This mutator should be paired with \code{hot_deck_lead_sampler} for use in CV.
It is applied to the training data after the train-test split,
to avoid data leakage.
}
