% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cv_crps}
\alias{plot_cv_crps}
\title{Plot CRPS of CV}
\usage{
plot_cv_crps(arg_list, cv_crps_out, ymax = NULL)
}
\arguments{
\item{arg_list}{list. The \verb{$arg_list} element in grid search output.}

\item{cv_crps_out}{The output of \code{\link[=calc_cv_crps]{calc_cv_crps()}} applied to the relevant
grid search output's element.}

\item{ymax}{optional numeric. The plots' y max.}
}
\description{
Plot the CRPS'es of cross validation output.
}
