% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_local_rows}
\alias{sample_local_rows}
\title{Sample a random row from local rows.}
\usage{
sample_local_rows(
  local_rows,
  .observation,
  current_obs,
  n_closest,
  derived_col_name
)
}
\arguments{
\item{local_rows}{tibble (NOT tsibble) The \code{local_rows} in \code{simulate_sample_path}.}

\item{.observation}{The observation column.}

\item{current_obs}{The \code{current_obs} in \code{simulate_sample_path}.}

\item{n_closest}{Scalar.}

\item{derived_col_name}{string. The name of the column that you want to
filter any NAs from.}
}
\value{
A row from \code{local_rows}, without the \code{offset} column anymore.
}
\description{
Procedure outline:
\itemize{
\item Filter rows with NAs in derived_col
\item Take the \code{n_closest} closest rows
\item Randomly sample one of these.
}
}
\details{
The \code{derived_col_name} may be something like the column of leads or
differences, or it could be the observation column itself.

TODO: make \code{derived_col_name} optional?

The returned row has its \code{offset} column removed. This is just easier for
this function; it would not be difficult to include, but I don't see a reason
to include it; its purpose should have been fulfilled within this function.
}
